Add TLS functionality to application
Extended config to allow configuration of TLS settings. In addition to
serving content directly over HTTPS, a middleware was added that detects
the `X-Forwarded-Proto` header and adds the HSTS header when the value
is set to `https`, which means that a reverse proxy is providing HTTPS
externally. Only private IP addresses can set this header. Setting the
HSTS header can be disabled with a separate config flag. Visiting the
server via HTTP although HTTPS is enabled is allowed but will result in
a redirect to HTTPS.