Improve concurrency of Eselsohr
This introduces a new process that runs together with the webserver
process. It is responsible for persisting data. A request can use the
new write queue in the app env to provide events, which are setters for
the resource along with a value.

The persistence process monitors the write queue and spawns a worker for
each event it receives. It keeps track of which collections are
currently worked on, so if multiple requests work on the same
collection, updates are still done in sequence, to avoid race
conditions.

Actions that persist data can collect all events and submit them at
once. This guarantees that all updates happens at once. Else, if the
application receives a lot of concurrent requests and requests submit a
lot of single updates, they would have to wait until all their updates
were run and this many open connections could exhaust the server.

The updated save function works now with the list of events, loads the
resource content only once, apply all updates and then writes the new
resource to the file.