Get rid of the system collection
The endpoint was public anyway, because the accesstoken for it was
hardcoded. Even if it was random, unauthorized people could just get it
from the start page. By removing it and making the endpoint for it
explicit, we gain some advantages:

* There is no CapOnlyResource anymore, which makes the code in the
repository simpler.
* We do not need the 0000-000... file anymore with the single
CreateResource token.
* A huge number of concurrent requests that want to create a new
resource could lead to too many open file handles on the system
collection. This problem was solved.