Make repository update functions thread-safe
The old update functions were not thread-safe. Although requests for a single
resource are handled sequentially, updates could read the same old
values in parallel. So if ten requests try to add something to a list,
only one new entry would be added.

The new functions don’t read the old value but expect it as an argument
to be passed in later. This happens during the fold in the `save`
function. This happens sequentially and guarantees that always the
newest value is used.