Do not use MonadUnliftIO as base class for effects
Using MonadUnliftIO as a base class makes it possible to use async
concurrent functions anywhere in the server, but also allows the use of
`liftIO`.
We want every effect to be wrapped so that no `liftIO` is used within
the business logic. Maybe this will change in the future or we just
check for `liftIO` during code reviews while allowing MonadUnliftIO for
concurrency features.