Get rid of the Frontend Action hack
Using POST forms for everything in the web client was not great for
usability. Every query is now send over GET with normal anchor tags.
This also simplifies the logic in the creation of items, because we
don’t have to manage Command, Query, and Frontend actions all at once.

With this change, the implementation of the sharing feature should also
be possible, because we can change the goto url more dynamically instead
of hard-coding it into the Frontend action token.