Add Nix support to the project
Based on this template:
https://github.com/srid/haskell-template/tree/6c8b0f92b8e62ad06b8ec1cebee06d44b8ea335a

This also includes some additions to the .vscode folder. A new bin/
folder includes quick commands to format everything in the project and
to run the project with ghcid.

Because Nix has some problems building fully static executables, we set
dynamically-linked builds as default and only enable the static builds
in the CI.