Improve test suite
Add unit, property-based, mocked, and integration tests.

Mocked tests use an in-memory collection instead of using serialized
files. Also scraped article titles are just the urls.

Integration tests use the real implementations. The collections are
stored in a temporary directory and a second Eselsohr instance is
started for fetching the title of the start page.